+++ src/aiosmtplib/api.py
@@ -1,6 +1,8 @@
 """
 Main public API.
 """
+import os
+import socket
 import ssl
 from email.message import Message
 from typing import Dict, List, Optional, Sequence, Tuple, Union, overload
@@ -17,7 +19,7 @@ async def send(
     message: Message,
     sender: Optional[str] = ...,
     recipients: Optional[Union[str, Sequence[str]]] = ...,
-    hostname: Optional[str] = ...,
+    hostname: str = ...,
     port: Optional[int] = ...,
     username: Optional[str] = ...,
     password: Optional[str] = ...,
@@ -32,6 +34,8 @@ async def send(
     client_key: Optional[str] = ...,
     tls_context: Optional[ssl.SSLContext] = ...,
     cert_bundle: Optional[str] = ...,
+    socket_path: None = ...,
+    sock: None = ...,
 ) -> Tuple[Dict[str, SMTPResponse], str]:
     ...
 
@@ -41,7 +45,7 @@ async def send(
     message: Union[str, bytes],
     sender: str = ...,
     recipients: Union[str, Sequence[str]] = ...,
-    hostname: Optional[str] = ...,
+    hostname: str = ...,
     port: Optional[int] = ...,
     username: Optional[str] = ...,
     password: Optional[str] = ...,
@@ -56,21 +60,117 @@ async def send(
     client_key: Optional[str] = ...,
     tls_context: Optional[ssl.SSLContext] = ...,
     cert_bundle: Optional[str] = ...,
+    socket_path: None = ...,
+    sock: None = ...,
 ) -> Tuple[Dict[str, SMTPResponse], str]:
     ...
 
 
-async def send(  # NOQA: F811
-    message,
-    sender=None,
-    recipients=None,
-    username=None,
-    password=None,
-    start_tls=False,
-    port=None,
-    use_tls=False,
-    **kwargs
-):
+@overload  # NOQA: F811
+async def send(
+    message: Message,
+    sender: Optional[str] = ...,
+    recipients: Optional[Union[str, Sequence[str]]] = ...,
+    hostname: None = ...,
+    port: None = ...,
+    username: Optional[str] = ...,
+    password: Optional[str] = ...,
+    mail_options: Optional[List[str]] = ...,
+    rcpt_options: Optional[List[str]] = ...,
+    timeout: Optional[float] = ...,
+    source_address: Optional[str] = ...,
+    use_tls: bool = ...,
+    start_tls: bool = ...,
+    validate_certs: bool = ...,
+    client_cert: Optional[str] = ...,
+    client_key: Optional[str] = ...,
+    tls_context: Optional[ssl.SSLContext] = ...,
+    cert_bundle: Optional[str] = ...,
+    socket_path: Union[str, bytes, os.PathLike] = ...,
+    sock: None = ...,
+) -> Tuple[Dict[str, SMTPResponse], str]:
+    ...
+
+
+@overload  # NOQA: F811
+async def send(
+    message: Union[str, bytes],
+    sender: str = ...,
+    recipients: Union[str, Sequence[str]] = ...,
+    hostname: None = ...,
+    port: None = ...,
+    username: Optional[str] = ...,
+    password: Optional[str] = ...,
+    mail_options: Optional[List[str]] = ...,
+    rcpt_options: Optional[List[str]] = ...,
+    timeout: Optional[float] = ...,
+    source_address: Optional[str] = ...,
+    use_tls: bool = ...,
+    start_tls: bool = ...,
+    validate_certs: bool = ...,
+    client_cert: Optional[str] = ...,
+    client_key: Optional[str] = ...,
+    tls_context: Optional[ssl.SSLContext] = ...,
+    cert_bundle: Optional[str] = ...,
+    socket_path: Union[str, bytes, os.PathLike] = ...,
+    sock: None = ...,
+) -> Tuple[Dict[str, SMTPResponse], str]:
+    ...
+
+
+@overload  # NOQA: F811
+async def send(
+    message: Message,
+    sender: Optional[str] = ...,
+    recipients: Optional[Union[str, Sequence[str]]] = ...,
+    hostname: None = ...,
+    port: None = ...,
+    username: Optional[str] = ...,
+    password: Optional[str] = ...,
+    mail_options: Optional[List[str]] = ...,
+    rcpt_options: Optional[List[str]] = ...,
+    timeout: Optional[float] = ...,
+    source_address: Optional[str] = ...,
+    use_tls: bool = ...,
+    start_tls: bool = ...,
+    validate_certs: bool = ...,
+    client_cert: Optional[str] = ...,
+    client_key: Optional[str] = ...,
+    tls_context: Optional[ssl.SSLContext] = ...,
+    cert_bundle: Optional[str] = ...,
+    socket_path: None = ...,
+    sock: socket.socket = ...,
+) -> Tuple[Dict[str, SMTPResponse], str]:
+    ...
+
+
+@overload  # NOQA: F811
+async def send(
+    message: Union[str, bytes],
+    sender: str = ...,
+    recipients: Union[str, Sequence[str]] = ...,
+    hostname: None = ...,
+    port: None = ...,
+    username: Optional[str] = ...,
+    password: Optional[str] = ...,
+    mail_options: Optional[List[str]] = ...,
+    rcpt_options: Optional[List[str]] = ...,
+    timeout: Optional[float] = ...,
+    source_address: Optional[str] = ...,
+    use_tls: bool = ...,
+    start_tls: bool = ...,
+    validate_certs: bool = ...,
+    client_cert: Optional[str] = ...,
+    client_key: Optional[str] = ...,
+    tls_context: Optional[ssl.SSLContext] = ...,
+    cert_bundle: Optional[str] = ...,
+    socket_path: None = ...,
+    sock: socket.socket = ...,
+) -> Tuple[Dict[str, SMTPResponse], str]:
+    ...
+
+
+async def send(message, sender=None, recipients=None, **kwargs):  # NOQA: F811
     """
     Send an email message. On await, connects to the SMTP server using the details
     provided, sends the message, then disconnects.
@@ -105,6 +205,10 @@ async def send(  # NOQA: F811
     :keyword tls_context: An existing :py:class:`ssl.SSLContext`, for TLS.
         Mutually exclusive with ``client_cert``/``client_key``.
     :keyword cert_bundle: Path to certificate bundle, for TLS verification.
+    :keyword socket_path: Path to a Unix domain socket. Not compatible with
+        hostname or port.
+    :keyword sock: An existing, connected socket object. If given, none of
+        hostname, port, or socket_path should be provided.
 
     :raises ValueError: required arguments missing or mutually exclusive options
         provided
@@ -115,14 +219,7 @@ async def send(  # NOQA: F811
         if not sender:
             raise ValueError("Sender must be provided with raw messages.")
 
-    client = SMTP(
-        port=port,
-        username=username,
-        password=password,
-        use_tls=use_tls,
-        start_tls=start_tls,
-        **kwargs
-    )
+    client = SMTP(**kwargs)
 
     async with client:
         if isinstance(message, Message):